Fix compilation warning with strlen function complete

+++ src/stdio/stdio.c
@@ -72,7 +72,7 @@ void clear() {
 
 void printf(const char *const str) {
     /* Printf! What more can we say :-) Happy days. */
-    for (int i = 0; i < strlen(str); ++i) {
+    for (size_t i = 0; i < strlen(str); ++i) {
         putchar(str[i], FONT_COLOR);
     }
 }
